yolo-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
ultralytics==8.3.57
Pillow>=10.0.0
pydantic>=2.0.0
